-Merge from trunk into rend32. -Fix an actor position interpolation bug.
SVN r3982 (rend32)

+++ client/src/r_plane.cpp
@@ -371,7 +371,7 @@ visplane_t* R_CheckPlane(visplane_t* pl, int start, int stop)
 		intrh = stop;
 	}
 
-	for (x = intrl ; x <= intrh && pl->top[x] == viewheight; x++)
+	for (x = intrl ; x <= intrh && pl->top[x] == (unsigned int)viewheight; x++)
 		;
 
 	if (x > intrh)
